requirements.txt
python-engineio==3.14.2
python-socketio[client]==4.6.0
